+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add a `warnings_as_errors` annotation processor option to enable warnings as errors. Downstream annotation processors should use the warning helper methods to ensure this option is used.
+
 ### [v0.66](https://github.com/realityforge/proton/tree/v0.66) (2026-04-17) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.65...v0.66)
 
 Changes in this release: